% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_github_token.R
\name{check_for_github_token}
\alias{check_for_github_token}
\alias{set_up_github_token}
\title{Verify a Github Token is Available}
\usage{
check_for_github_token(token = NULL)
}
\arguments{
\item{token}{Character.  A string giving the token.}
}
\value{
Logical.  \code{TRUE} if the token is found in the workspace.
}
\description{
You will need to set up an access token at Github to avoid severe
limits to the rate at which you can search.
}
\section{Obtaining a Token}{

One first needs to create an "app" at Github following the instructions
\href{https://github.com/settings/developers}{here}.
Then do the following steps:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library("httr")
myapp <- oauth_app("app_name", key = "app_key", secret = "app_secret")
github_token <- oauth2.0_token(oauth_endpoints("github"), myapp)
}\if{html}{\out{</div>}}
}

\author{
Bryan A. Hanson
}
