% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_page_date.R
\name{find_page_date}
\alias{find_page_date}
\title{Find Date on a Web Page}
\usage{
find_page_date(string)
}
\arguments{
\item{string}{Character.  A single character string containing the contents of a web page.}
}
\value{
YMD in ISO 8601 format.  Never returns a date in the future.
}
\description{
Find the most recent date on a web page that is not in the future.
Looking for anything that can be converted to YMD.
Inspired by the Python package htmldate, but no code in common.
}
\examples{
res <- find_page_date(flatten_web_page("http://dirk.eddelbuettel.com/cranberries/cran/new/"))

}
