% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_web_page.R
\name{flatten_web_page}
\alias{flatten_web_page}
\title{Grab a Web Page and Flatten to a Simple Character Vector}
\usage{
flatten_web_page(url)
}
\arguments{
\item{url}{Valid URL string.  User should check validity before calling this function.}
}
\value{
A character vector containing the contents of the web page (xml is stripped out).
Lines containing only whitespace + newlines are removed as well.
}
\description{
Convert a web page to a simple character vector which is fast to search.
If looking for something specific use the methods in \code{rvest}.
}
\examples{
url <- "https://chemospec.org/posts/2020-01-01-Intro-F4S/2021-01-01-Intro-F4S.html"
res <- flatten_web_page(url)

}
