% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_dates.R
\name{get_github_dates}
\alias{get_github_dates}
\title{Get the Most Recent Commit or Issue Date from a Github Repository}
\usage{
get_github_dates(url, what = "commits")
}
\arguments{
\item{url}{Character.  A string giving the URL of a Github repository.}

\item{what}{Character. Either \code{commits} or \code{issues}.}
}
\value{
A date in Y-M-D format (ISO 8601 essentially) and class \code{Date},
unless there is a problem, in which case \code{NA}.
}
\description{
Access a Github repository and retrieve the date of the most recent commit or issue.
}
\examples{
tst <- get_github_dates("https://github.com/bryanhanson/ChemoSpec")

}
\author{
Bryan A. Hanson
}
