% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good_url.R
\name{good_url}
\alias{good_url}
\title{Verify a URL}
\usage{
good_url(url, quiet = TRUE, ...)
}
\arguments{
\item{url}{Character.  A string giving the URL.}

\item{quiet}{Logical.  If \code{TRUE} stay quiet.}

\item{...}{Extra arguments to be passed to \code{httr::GET}.}
}
\value{
A logical; \code{TRUE} if the URL was accessible.
}
\description{
This function will attempt to access a URL, and if successful, return \code{TRUE}.
Success is defined as getting a 200 status code.
}
\examples{
good <- good_url("https://example.com")

}
\author{
Bryan A. Hanson
}
