% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_github_for_topics.R
\name{search_github_for_topics}
\alias{search_github_for_topics}
\title{Search Github Repos for Selected Topics}
\usage{
search_github_for_topics(
  topics = NULL,
  browse = FALSE,
  username = NULL,
  token = NULL
)
}
\arguments{
\item{topics}{Character.  A vector of topics to search.}

\item{browse}{Logical.  Shall the repos found be opened in a browser?}

\item{username}{Character.  Github username.}

\item{token}{Character.  Github authentification token.}
}
\value{
Invisibly, a data frame containing the repo data.
}
\description{
Search Github repositories for selected topics.  You will need to
set up an access token at Github to avoid severe
limits to the rate at which you can search.
}
\author{
Bryan A. Hanson
}
