% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pypi_for_topics.R
\name{search_pypi_for_topics}
\alias{search_pypi_for_topics}
\title{Search PyPi for Selected Topics}
\usage{
search_pypi_for_topics(topics = NULL, browse = FALSE)
}
\arguments{
\item{topics}{Character.  A vector of topics to search.}

\item{browse}{Logical.  Shall the repos found be opened in a browser?
This could be a lot of tabs!}
}
\value{
Invisibly, a data frame containing the repo data.  Elements are:
\itemize{
\item name  Name of the project.
\item description Short summary of the project.
\item repository The url to the repository on PyPi.org.
\item author_email E-mail of the author.
\item maintainer_email E-mail of the maintainer.
}
}
\description{
Search PyPi.org for selected topics.  PyPi.org does not require authentification.
}
\examples{
\dontrun{
# This returns less than 10 pages
res <- search_pypi_for_topics("narwhal", browse = TRUE)
}

}
\author{
Bryan A. Hanson
}
