% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webutils-package.R
\docType{package}
\name{webu-package}
\alias{webu-package}
\alias{webu}
\title{Functions for Accessing the Web}
\description{
Functions useful for accessing Github and PyPi via the API, along with
a few other functions helpful for working on the web.
}
\author{
Bryan A. Hanson
}
\keyword{package}
